Refactor : 예외 처리 생성자 수정
