+++ src/main/java/com/ee06/wooms/domain/users/exception/ex/UserEmailNotFoundException.java
@@ -2,10 +2,14 @@ package com.ee06.wooms.domain.users.exception.ex;
 
 import com.ee06.wooms.global.exception.ErrorCode;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+
+import static com.ee06.wooms.global.exception.ErrorCode.NOT_FOUND_EMAIL_USER;
 
 @Getter
-@RequiredArgsConstructor
 public class UserEmailNotFoundException extends RuntimeException {
-    final ErrorCode errorCode;
+    private final ErrorCode errorCode;
+
+    public UserEmailNotFoundException() {
+        this.errorCode = NOT_FOUND_EMAIL_USER;
+    }
 }
