+++ src/main/java/com/ee06/wooms/domain/users/exception/ex/UserExistException.java
@@ -2,10 +2,14 @@ package com.ee06.wooms.domain.users.exception.ex;
 
 import com.ee06.wooms.global.exception.ErrorCode;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+
+import static com.ee06.wooms.global.exception.ErrorCode.EXIST_USER;
 
 @Getter
-@RequiredArgsConstructor
 public class UserExistException extends RuntimeException {
     final ErrorCode errorCode;
+
+    public UserExistException() {
+        this.errorCode = EXIST_USER;
+    }
 }
