+++ src/main/java/com/ee06/wooms/domain/users/exception/ex/UserNotSentEmailException.java
@@ -2,10 +2,14 @@ package com.ee06.wooms.domain.users.exception.ex;
 
 import com.ee06.wooms.global.exception.ErrorCode;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+
+import static com.ee06.wooms.global.exception.ErrorCode.NOT_SENT_EMAIL_USER;
 
 @Getter
-@RequiredArgsConstructor
 public class UserNotSentEmailException extends RuntimeException {
     private final ErrorCode errorCode;
+
+    public UserNotSentEmailException() {
+        this.errorCode = NOT_SENT_EMAIL_USER;
+    }
 }
