+++ src/main/java/com/ee06/wooms/domain/users/service/CustomOAuth2UserService.java
@@ -8,7 +8,6 @@ import com.ee06.wooms.domain.users.entity.SocialProvider;
 import com.ee06.wooms.domain.users.entity.User;
 import com.ee06.wooms.domain.users.entity.UserStatus;
 import com.ee06.wooms.domain.users.repository.UserRepository;
-import com.ee06.wooms.global.exception.ErrorCode;
 import com.ee06.wooms.global.oauth.exception.NotFoundPlatformException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -34,7 +33,7 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
         OAuth2Response oAuth2Response =
                 provideOAuth(userRequest.getClientRegistration().getRegistrationId(), oAuth2User)
-                .orElseThrow(() -> new NotFoundPlatformException(ErrorCode.NOT_FOUND_PLATFORM_SERVICE));
+                .orElseThrow(NotFoundPlatformException::new);
 
         String userEmail = oAuth2Response.getEmail();
