+++ src/main/java/com/ee06/wooms/global/jwt/exception/InvalidTokenException.java
@@ -2,10 +2,14 @@ package com.ee06.wooms.global.jwt.exception;
 
 import com.ee06.wooms.global.exception.ErrorCode;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+
+import static com.ee06.wooms.global.exception.ErrorCode.INVALID_TOKEN;
 
 @Getter
-@RequiredArgsConstructor
 public class InvalidTokenException extends RuntimeException {
     private final ErrorCode errorCode;
+
+    public InvalidTokenException() {
+        this.errorCode = INVALID_TOKEN;
+    }
 }
